+++ admin/src/js/main.js
@@ -79,7 +79,6 @@ require('../../../webapp/src/js/services/changes');
 require('../../../webapp/src/js/services/contact-muted');
 require('../../../webapp/src/js/services/contact-schema');
 require('../../../webapp/src/js/services/db');
-require('../../../webapp/src/js/services/download-url');
 require('../../../webapp/src/js/services/export');
 require('../../../webapp/src/js/services/extract-lineage');
 require('../../../webapp/src/js/services/file-reader');
