+++ admin/src/templates/export_contacts.html
@@ -6,7 +6,6 @@
       <i class="fa fa-arrow-down"></i>
       <span translate>Export</span>
     </a>
-    <span class="loader inline small" ng-show="exporting"></span>
   </div>
 
   <div class="section">
