+++ admin/src/templates/export_feedback.html
@@ -4,7 +4,6 @@
       <i class="fa fa-arrow-down"></i>
       <span translate>Export</span>
     </a>
-    <span class="loader inline small" ng-show="exporting"></span>
   </div>
   <div class="section">
     <h3 translate translate-values="feedback.page">Showing number of total</h3>
