+++ admin/src/templates/export_messages.html
@@ -4,6 +4,5 @@
       <i class="fa fa-arrow-down"></i>
       <span translate>Export</span>
     </a>
-    <span class="loader inline small" ng-show="exporting"></span>
   </div>
 </div>
