+++ webapp/src/js/services/export.js
@@ -1,21 +1,19 @@
-var ajaxDownload = require('../modules/ajax-download');
+const ajaxDownload = require('../modules/ajax-download');
+const KNOWN_TYPES = ['reports', 'contacts', 'messages', 'feedback'];
 
 angular.module('inboxServices').factory('Export',
   function(
-    $http,
-    $log,
-    DownloadUrl
+    $log
   ) {
     'ngInject';
     'use strict';
-    return function(filters, typeName) {
-      return DownloadUrl(filters, typeName)
-        .then(function(url) {
-          ajaxDownload.download(url);
-        })
-        .catch(function(err) {
-          $log.error('Error exporting ' + typeName, err);
-        });
+    return function(type, filters) {
+      if (!KNOWN_TYPES.includes(type)) {
+        return $log.error(new Error('Unknown download type: ' + type));
+      }
+      const params = filters ? '?' + $.param({ filters: filters }) : '';
+      const url = '/api/v2/export/' + type + params;
+      ajaxDownload.download(url);
     };
   }
 );
